+++ TVRename/Custom Controls/ListViewFlickerFree.cs
@@ -5,6 +5,7 @@
 // 
 // Copyright (c) TV Rename. This code is released under GPLv3 https://github.com/TV-Rename/tvrename/blob/master/LICENSE.md
 // 
+
 namespace TVRename
 {
     using System.Windows.Forms;
