+++ TVRename/Custom Controls/MyListView.cs
@@ -6,10 +6,8 @@
 // Copyright (c) TV Rename. This code is released under GPLv3 https://github.com/TV-Rename/tvrename/blob/master/LICENSE.md
 // 
 
-using System;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
-using JetBrains.Annotations;
+using TVRename.Utility;
 
 // Starting from:
 // http://social.msdn.microsoft.com/Forums/ja-JP/csharpexpressja/thread/67475927-015c-4206-b5e7-d67504edb3a1
@@ -97,55 +95,14 @@ namespace TVRename
             }
         }
 
-        // The 'TopItem' function doesn't work in a ListView if groups are enabled. This is meant to be a workaround.
-        // Problem is, it just doesn't work and I don't know why!
-        // ReSharper disable once UnusedMember.Local
-        private const int SB_HORZ = 0;
-        private const int SB_VERT = 1;
-        private const int LVM_FIRST = 0x1000;
-        private const int LVM_SCROLL = LVM_FIRST + 20;
-
-        public int GetScrollVerticalPos() => NativeMethods.GetScrollPos(Handle, SB_VERT);
-
         public void SetScrollVerticalPos(int position)
         {
-            int currentPos = NativeMethods.GetScrollPos(Handle, SB_VERT);
-            int delta = -(currentPos - position);
-            NativeMethods.SendMessage(Handle, LVM_SCROLL, IntPtr.Zero, (IntPtr)delta); // First param is horizontal scroll amount, second is vertical scroll amount
+            ListViewNativeMethods.SetScrollVerticalPos(this,position);
         }
 
-        public void HideCheckbox([NotNull] ListViewItem item)
+        public int GetScrollVerticalPos()
         {
-            Lvitem lviItem = new Lvitem
-            {
-                iItem = item.Index, mask = LVIF_STATE, stateMask = LVIS_STATEIMAGEMASK, state = 0
-            };
-
-            SendMessage(Handle, LVM_SETITEM, IntPtr.Zero, ref lviItem);
+            return ListViewNativeMethods.GetScrollVerticalPos(this);
         }
-
-        private const int LVIF_STATE = 0x8;
-        private const int LVIS_STATEIMAGEMASK = 0xF000;
-        private const int LVM_SETITEM = LVM_FIRST + 76;
-
-        // suppress warnings for interop
-#pragma warning disable 0649
-        private struct Lvitem
-        {
-            public int mask;
-            public int iItem;
-            public int iSubItem;
-            public int state;
-            public int stateMask;
-            [MarshalAs(UnmanagedType.LPTStr)]
-            public String lpszText;
-            public int cchTextMax;
-            public int iImage;
-            public IntPtr iParam;
-        }
-#pragma warning restore 0649
-
-        [DllImport("user32.dll")]
-        private static extern IntPtr SendMessage(IntPtr hWnd, uint msg, IntPtr wParam, ref Lvitem lParam);
     }
 }
